// parking_admin/src/app/models/user.ts
export class User {
    id: number;
    firstName: string = "";
    lastName: string = "";
    email: string = "";
    phoneNumber: string = "";
    username: string = "";
    password: string = "";
    active: string = "";
    lockUser: string = "";
    userGroups: string = "";
    files: string = "";
}